Continued moosification!
Starting to move Player objectes to typed moose system.   Starting
to examine the world object and how I can improve the data model there too.

Added code to produce a 'dot' file from the defined world, so we can use the
dot utility to create a dungeon map.  It already found a few inconsistencies in
the world, that were corrected. This update includes .pdf of map.